Moved Settings and updated SettingsTest
Settings moved to their own folder as per #70.
SettingsTest points to new location of Settings.
