+++ Test/SettingsTest.cpp
@@ -1,4 +1,4 @@
-#include "../Source Files/Utility/Settings.h"
+#include "../Source Files/Utility/Settings/Settings.h"
 #include <cassert>
 #include <iostream>
